requirements.txt
numpy
pandas
scikit-learn
xgboost
imbalanced-learn
shap
matplotlib
fastapi
uvicorn
pydantic
streamlit
requests
joblib
mlflow
google-generativeai
